Add "colour.colorimetry.generation" module and move accordingly "colour.constant_spd", "colour.zeros_spd" and  "colour.ones_spd" definitions.
