+++ colour/colorimetry/__init__.py
@@ -2,14 +2,15 @@
 
 from __future__ import absolute_import
 
-from .spectrum import (SpectralShape, SpectralPowerDistribution,
-                       MultiSpectralPowerDistribution, DEFAULT_SPECTRAL_SHAPE,
-                       constant_spd, zeros_spd, ones_spd)
+from .spectrum import (SpectralShape, DEFAULT_SPECTRAL_SHAPE,
+                       SpectralPowerDistribution,
+                       MultiSpectralPowerDistribution)
 from .blackbody import (blackbody_spd, blackbody_spectral_radiance, planck_law)
 from .cmfs import (LMS_ConeFundamentals, RGB_ColourMatchingFunctions,
                    XYZ_ColourMatchingFunctions)
 from .dataset import *  # noqa
 from . import dataset
+from .generation import constant_spd, zeros_spd, ones_spd
 from .tristimulus import (SPECTRAL_TO_XYZ_METHODS,
                           MULTI_SPECTRAL_TO_XYZ_METHODS)
 from .tristimulus import spectral_to_XYZ, multi_spectral_to_XYZ
@@ -56,9 +57,8 @@ from .yellowness import yellowness
 from .yellowness import yellowness_ASTMD1925, yellowness_ASTME313
 
 __all__ = [
-    'SpectralShape', 'SpectralPowerDistribution',
-    'MultiSpectralPowerDistribution', 'DEFAULT_SPECTRAL_SHAPE', 'constant_spd',
-    'zeros_spd', 'ones_spd'
+    'SpectralShape', 'DEFAULT_SPECTRAL_SHAPE', 'SpectralPowerDistribution',
+    'MultiSpectralPowerDistribution'
 ]
 __all__ += ['blackbody_spd', 'blackbody_spectral_radiance', 'planck_law']
 __all__ += [
@@ -66,6 +66,7 @@ __all__ += [
     'XYZ_ColourMatchingFunctions'
 ]
 __all__ += dataset.__all__
+__all__ += ['constant_spd', 'zeros_spd', 'ones_spd']
 __all__ += ['SPECTRAL_TO_XYZ_METHODS', 'MULTI_SPECTRAL_TO_XYZ_METHODS']
 __all__ += ['spectral_to_XYZ', 'multi_spectral_to_XYZ']
 __all__ += [
