+++ colour/colorimetry/tests/test_spectrum.py
@@ -11,8 +11,7 @@ import scipy
 from distutils.version import LooseVersion
 
 from colour.colorimetry.spectrum import (
-    SpectralShape, SpectralPowerDistribution, MultiSpectralPowerDistribution,
-    constant_spd, zeros_spd, ones_spd)
+    SpectralShape, SpectralPowerDistribution, MultiSpectralPowerDistribution)
 from colour.utilities import tstack
 
 __author__ = 'Colour Developers'
@@ -27,8 +26,7 @@ __all__ = [
     'INTERPOLATED_SAMPLE_SPD_DATA', 'INTERPOLATED_NON_UNIFORM_SAMPLE_SPD_DATA',
     'NORMALISED_SAMPLE_SPD_DATA', 'CIE_1931_2_DEGREE_STANDARD_OBSERVER',
     'CMFS_DATA', 'TestSpectralShape', 'TestSpectralPowerDistribution',
-    'TestMultiSpectralPowerDistribution', 'TestConstantSpd', 'TestZerosSpd',
-    'TestOnes_spd'
+    'TestMultiSpectralPowerDistribution'
 ]
 
 SAMPLE_SPD_DATA = {
@@ -1707,69 +1705,6 @@ MultiSpectralPowerDistribution.to_spds` method.
             self.assertEqual(spd.strict_name, '{0} - {1}'.format(
                 self._strict_labels[i],
                 self._non_uniform_sample_multi_spd.name))
-
-
-class TestConstantSpd(unittest.TestCase):
-    """
-    Defines :func:`colour.colorimetry.spectrum.constant_spd` definition unit
-    tests methods.
-    """
-
-    def test_constant_spd(self):
-        """
-        Tests :func:`colour.colorimetry.spectrum.constant_spd`
-        definition.
-        """
-
-        spd = constant_spd(np.pi)
-
-        self.assertAlmostEqual(spd[360], np.pi, places=7)
-
-        self.assertAlmostEqual(spd[555], np.pi, places=7)
-
-        self.assertAlmostEqual(spd[780], np.pi, places=7)
-
-
-class TestZerosSpd(unittest.TestCase):
-    """
-    Defines :func:`colour.colorimetry.spectrum.zeros_spd` definition unit
-    tests methods.
-    """
-
-    def test_zeros_spd(self):
-        """
-        Tests :func:`colour.colorimetry.spectrum.zeros_spd`
-        definition.
-        """
-
-        spd = zeros_spd()
-
-        self.assertEqual(spd[360], 0)
-
-        self.assertEqual(spd[555], 0)
-
-        self.assertEqual(spd[780], 0)
-
-
-class TestOnes_spd(unittest.TestCase):
-    """
-    Defines :func:`colour.colorimetry.spectrum.ones_spd` definition unit
-    tests methods.
-    """
-
-    def test_ones_spd(self):
-        """
-        Tests :func:`colour.colorimetry.spectrum.ones_spd`
-        definition.
-        """
-
-        spd = ones_spd()
-
-        self.assertEqual(spd[360], 1)
-
-        self.assertEqual(spd[555], 1)
-
-        self.assertEqual(spd[780], 1)
 
 
 if __name__ == '__main__':
